# [POC] Faster channel switching