|  | before | after |
| --- |--- | --- |
|channel_info| 0.170s | 0.04s |

cakal som ze to bude vacsia uspora ~200ms na pomalsich spojeniach asi moc viditelne nebude

todo:
- [ ] check cookies lifetime

fixes:
https://github.com/Sorien/plugin.video.sl/issues/42