@Sorien 2019-07-23
hmm tie session data by nemalo byt treba zmazat lebo nazov toho suboru sa zmenil ... a mam pocit, ze cez toto nove api kodi prepina kanaly rychlesie dokonca mozno rychlejsie ako televizor :)